// src/types/types.ts
export type MenuType = {
    id: number;
    slug: string;
    title: string;
    desc?: string;
    img?: string;
    color: string;
  }[];

export type ProductType = {
    id: number;
    title: string;
    desc?: string;
    cat: string;
    img?: string;
    price: number;
    options?: { title: string; additionalPrice: number }[];
  };